Backend/requirements.txt
absl-py==2.2.1
astunparse==1.6.3
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
firebase-admin==6.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
keras==3.9.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
msgpack==1.1.0
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
rich==14.0.0
rsa==4.9
scipy==1.15.2
sentencepiece==0.2.0
setuptools==78.1.0
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.0
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0